[techTask]Repository: Move MovieRepository to the right package
